Make ApplicationRegistrator an interface
closes #1428
